api/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
bangla==0.0.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.38.12
botocore==1.38.12
braceexpand==0.1.7
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.0
cytoolz==1.0.1
datasets==3.6.0
dateparser==1.1.8
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fiddle==0.3.0
filelock==3.13.1
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.6.1
funcy==2.0
future==1.0.0
g2p-en==2.1.0
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.0
huggingface-hub==0.31.1
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
intervaltree==3.1.0
ipython==8.36.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
jiwer==3.1.0
jmespath==1.0.1
joblib==1.5.0
jsonlines==1.2.0
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.31.0
libcst==1.7.0
librosa==0.10.0
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.43.0
loguru==0.7.3
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==4.0.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
nemo-toolkit==2.2.1
networkx==3.3
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
omegaconf==2.3.0
onnx==1.17.0
optuna==4.3.0
packaging==24.2
pandas==1.5.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.12.0
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
SudachiDict-core==20250129
SudachiPy==0.6.10
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
texterrors==0.5.1
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.7.1
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.48.3
TTS==0.22.0
typeguard==4.4.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.19.11
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.111
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
